Create gui_interface.py
Start of the GUI file. Goal is to provide interface for non-command line functionality. 
Issue is knowing what parameters the user has.

Going forward - Maybe have all the parameters listed and allow the user to enter what they know. Then we can calculate
what can be calculated and what parameters still need to be figured out or guesstimated on.